test(api): move vector-store signature tests out of generated files
